+++ fs-api/src-linux/System/FS/IO/Internal/Error.hs
@@ -1,0 +1,7 @@
+module System.FS.IO.Internal.Error (sameError) where
+
+import           System.FS.API.Types (FsError, sameFsError)
+
+sameError :: FsError -> FsError -> Bool
+sameError = sameFsError
+